+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 * 🔧 Upcoming/Recent: rename "Infinite calendar" option to the more applicable "Unlimited days".
 * 🔧 Widget: note general item limit in configuration screen, rename "Infinite calendar" option to
   the more applicable "Unlimited days", move hide watch button option to appearance category.
+* 🔧 On Android 16 on Samsung devices, increase the limit of widget items back to 100.
 
 ## Version 2025.2