+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 
 ## Version 2025.3
 
+### 2025.3.0 - 2025-10-24 🧪
+
 * 🔧 Upcoming/Recent: rename "Infinite calendar" option to the more applicable "Unlimited days".
 * 🔧 Widget: note general item limit in configuration screen, rename "Infinite calendar" option to
   the more applicable "Unlimited days", move hide watch button option to appearance category.